fix: change horas_planejadas type from int to float in DisponibilidadeProjetoDetalhe schema

+++ app/models/schemas.py
@@ -209,7 +209,7 @@ class ProjetoInfo(BaseSchema):
 
 class DisponibilidadeProjetoDetalhe(BaseModel):
     projeto: ProjetoInfo
-    horas_planejadas: int
+    horas_planejadas: float
 
 class DisponibilidadeMensal(BaseSchema):
     mes: int
